requirements.txt
click==7.1.2
inquirer==2.7.0
docker_py==1.10.6
PyYAML==5.4.1
